% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_data.R
\docType{data}
\name{dif_fut_sdm}
\alias{dif_fut_sdm}
\title{A map with the difference between a current and future projection for the distribution of a marine species.}
\format{
A RasterLayer with the map indicating the difference. Values are:
\describe{
  \item{0}{Unsuitable}
  \item{2}{Lost areas}
  \item{2}{Gain areas}
  \item{3}{Kept areas}
}
}
\source{
The author.
}
\usage{
dif_fut_sdm
}
\description{
This consists of a RasterLayer with the difference between current and future binarized projections for a SDM generated with BRT (Boosted Regression Trees) for the marine species.
}
\keyword{datasets}
