% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_data.R
\docType{data}
\name{pa_data}
\alias{pa_data}
\title{Presence/absence data for a marine species}
\format{
A data frame with 4400 rows and 3 variables:
\describe{
  \item{x}{Longitude}
  \item{y}{Latitude}
  \item{pa}{Indicates presence (1) or absence (0)}
}
}
\source{
The author.
}
\usage{
pa_data
}
\description{
A dataset containing presence and absence data for a marine species. This was extracted from an SDM for the coral species *Montastraea cavernosa*, but DOES NOT reflect the actual species distribution. You should think of it as a **virtual species** dataset.
}
\keyword{datasets}
