% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_highlight.R
\name{sdm_highlight}
\alias{sdm_highlight}
\alias{sdm_highlight,Raster-method}
\alias{sdm_highlight,data.frame-method}
\title{SDM Leaflet map highlighting extreme values}
\usage{
sdm_highlight(
  sdm,
  quantile,
  both.sides = FALSE,
  pts = NULL,
  pal = NULL,
  crs = "standard",
  cluster = FALSE,
  simplify = TRUE
)

\S4method{sdm_highlight}{Raster}(
  sdm,
  quantile,
  both.sides = FALSE,
  pts = NULL,
  pal = NULL,
  crs = "standard",
  cluster = FALSE,
  simplify = TRUE
)

\S4method{sdm_highlight}{data.frame}(
  sdm,
  quantile,
  both.sides = FALSE,
  pts = NULL,
  pal = NULL,
  crs = "standard",
  cluster = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{sdm}{The SDM/ENM result. Should be a data.frame (XYZ columns) or in the Raster* format.}

\item{quantile}{A single numeric value between 0 and 1 indicating which quantile to use (will be the same for both sides, if `both.sides` set to `TRUE`). For example, if you want the 25 percent higher values, you can set the quantile to 0.75.
If you want both higher and lower 25 percent, just put 0.75, and set `both.sides = TRUE`. Note: setting quantile 1 makes no sense, as it will be 100 percent. The same for setting it 0.5 and `both.sides = TRUE`.}

\item{both.sides}{If set to TRUE, than both the higher and lower values will be ploted.}

\item{pts}{A data frame containing the presence or presence/absence points (optional). The first column should be longitude (x) and the sencond latitude (y). In the case of presence/absence data, an additional collumn should be provided, coded as 0 (absence) and 1 (presence).}

\item{pal}{Character string indicating the name of the continuous mode palette (see \link[sdmvis]{gen_pal}). If not supplied, the default will be used.}

\item{crs}{Enables to change the default projection used in the Leaflet package. For now, not functional.}

\item{cluster}{Should the points be clustered (i.e., aggregated)? Only valid if `pts` is supplied. Default is FALSE.}

\item{simplify}{Should the polygons be simplified? If TRUE, the output became lighter.}
}
\value{
A Leaflet map.
}
\description{
Create a Leaflet map highlighting extreme values in an SDM output. This can either be minimum, maximum or both values set as a quantile.
}
\section{Functions}{
\itemize{
\item \code{sdm_highlight,Raster-method}: Method for Raster*

\item \code{sdm_highlight,data.frame-method}: Method for data frames
}}

\examples{
library(sdmvis)
# Load data
data("original_sdm")
data("pa_data")

# Plot
sdm_highlight(sdm = original_sdm[[1]],
                quantile = .75,
                pts = pa_data,
                both.sides = TRUE)

}
