% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmvis-package.R
\docType{package}
\name{sdmvis-package}
\alias{sdmvis}
\alias{sdmvis-package}
\title{sdmvis: Interactive tools for visualizing SDM/ENM results}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package contains functions for visualizing SDM or ENM outputs interactively.
}
\author{
\strong{Maintainer}: Silas C. Principe \email{silasprincipe@usp.br} (\href{https://orcid.org/0000-0002-8059-6304}{ORCID})

}
\keyword{internal}
