% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_leaflet.R
\name{simple_leaflet}
\alias{simple_leaflet}
\title{Simple SDM leaflet map}
\usage{
simple_leaflet(
  sdm,
  mode = "bin",
  pts = NULL,
  pal = NULL,
  layernames = NULL,
  crs = "standard"
)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
