% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_data.R
\docType{data}
\name{thresholded_sdm}
\alias{thresholded_sdm}
\title{A thresholded SDM for a marine species}
\format{
A RasterStack with two layers:
\describe{
  \item{current.map}{Projection for the current period.}
  \item{future.map}{Projection for the future period.}
}
}
\source{
The author.
}
\usage{
thresholded_sdm
}
\description{
This consists of a RasterStack of current and future projections for a SDM generated with BRT (Boosted Regression Trees) for the marine species using only two variables, but thresholded with a random value to produce binary maps.
}
\keyword{datasets}
