% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pal.R
\name{gen_pal}
\alias{gen_pal}
\title{Generate binary or 4 color palettes}
\usage{
gen_pal(mode, pal, type = "leaflet")
}
\arguments{
\item{mode}{Mode to generate the palette. "bin" is suitable for binary maps (thresholded maps). "quad" will return a 4 color palette, suitable for maps showing differences from a (binary) reference to a distinct scenario (either future or past).}

\item{pal}{Which palete to use. Character string indicating the name of the palette. Can also be a character vector of colors, in which case two or four should be supplied. The following pre-seted (and tested) palettes are available and are recommended:
\itemize{
 \item{"YeBl":}{ Yellow and Blue}
 \item{"GrBl":}{ Green and Blue}
}
All of these palettes are colorblind safe.}

\item{type}{Which type of palette to return. If "leaflet" (default), a leaflet ready palette is returned. Any other value will return a character vector of colors.}
}
\value{
A palette to produce interactive SDM/ENM maps.
}
\description{
This is an accessory function to generate binary or 4 color palettes to be used in the leaflet maps.
}
\examples{
library(raster)
library(leaflet)
library(rgdal)

sdm <- raster(nrows=10, ncols=10)
sdm[] <- sample(c(0, 1, NA), size = ncell(sdm), replace = TRUE)

binpal <- gen_pal(mode = "bin", pal = "YeBl")

leaflet()\%>\%
addProviderTiles("Esri.OceanBasemap") \%>\%
addRasterImage(sdm, colors = binpal, opacity = 0.9)
}
