% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_leaflet.R
\name{sdm_leaflet}
\alias{sdm_leaflet}
\title{SDM Leaflet map}
\usage{
sdm_leaflet(
  sdm,
  mode = "bin",
  pts = NULL,
  pal = NULL,
  layernames = NULL,
  crs = "standard"
)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{

}
