% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_thresh.R
\name{sdm_thresh}
\alias{sdm_thresh}
\alias{sdm_thresh,Raster-method}
\alias{sdm_thresh,data.frame-method}
\title{SDM Leaflet map with different thresholds}
\usage{
sdm_thresh(
  sdm,
  thresh,
  tname = NULL,
  pts = NULL,
  pal = NULL,
  crs = "standard",
  cluster = FALSE,
  simplify = TRUE,
  thresh.color = NULL
)

\S4method{sdm_thresh}{Raster}(
  sdm,
  thresh,
  tname = NULL,
  pts = NULL,
  pal = NULL,
  crs = "standard",
  cluster = FALSE,
  simplify = TRUE,
  thresh.color = NULL
)

\S4method{sdm_thresh}{data.frame}(
  sdm,
  thresh,
  tname = NULL,
  pts = NULL,
  pal = NULL,
  crs = "standard",
  cluster = FALSE,
  simplify = TRUE,
  thresh.color = NULL
)
}
\arguments{
\item{sdm}{The SDM/ENM result. Should be in the Raster* format.}

\item{thresh}{A vector containing the thresholds to be applied (numeric values).}

\item{tname}{An optional character vector containing the names of the thresholds to be used in the legend. If not supplied, numerals will be used instead.}

\item{pts}{A data frame containing the presence or presence/absence points (optional). The first column should be longitude (x) and the sencond latitude (y). In the case of presence/absence data, an additional collumn should be provided, coded as 0 (absence) and 1 (presence).}

\item{pal}{Character string indicating the name of the continuous mode palette (see \link[sdmvis]{gen_pal}). If not supplied, the default will be used.}

\item{crs}{Enables to change the default projection used in the Leaflet package. For now, not functional.}

\item{cluster}{Should the points be clustered (i.e., aggregated)? Only valid if `pts` is supplied. Default is FALSE.}

\item{simplify}{Should the polygons be simplified? If TRUE, the output became lighter.}

\item{thresh.color}{Vector of color(s) to be used for the threshold(s) polygon(s)}
}
\value{
A Leaflet map.
}
\description{
Create a Leaflet map with the results of an SDM and draw how different thresholds will result.
This map can be explored interactively in the viewer, so its possible to have a better comprehension of the outcomes of your SDM.
}
\section{Functions}{
\itemize{
\item \code{sdm_thresh,Raster-method}: Method for Raster*

\item \code{sdm_thresh,data.frame-method}: Method for data frames
}}

\examples{
library(sdmvis)
# Load data
data("original_sdm")
data("pa_data")

# Plot
sdm_thresh(sdm = original_sdm[[1]],
           thresh = -2,
           tname = "TSS",
           pts = pa_data)

}
