% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_report.R
\name{var_report}
\alias{var_report}
\title{Generate}
\usage{
var_report(var, pts, mode, type = "rmd")
}
\arguments{
\item{var}{A raster or data.frame with the environmental variables. In the case of a data.frame, the first two collumns should be, respectively, longitude and latitude.}

\item{pts}{A data frame containing the presence or presence/absence points (optional). The first column should be longitude (x) and the sencond latitude (y). In the case of presence/absence data, an additional collumn should be provided, coded as 0 (absence) and 1 (presence).}

\item{mode}{Which type of result return:
\itemize{
 \item{"full.summary"}{A summary of each variable.}
 \item{"pts.summary"}{A summary for each variable, extracted from the points.}
 \item{"d.plots"}{Stuff}
 \item{"pt.plots"}{Stuff}
 \item{"de.plots"}{Stuff}
}}

\item{type}{}
}
\value{
A Leaflet map and optionally an .html file with summary statistics.
}
\description{
Create a Leaflet map with the environmental variables.
This map can be explored interactively in the viewer.
}
\examples{
library(sdmvis)
# Load data
data("thresholded_sdm")
data("pa_data")
# Create a list of SDMs and names
sdm.list <- list(thresholded_sdm[[1]], thresholded_sdm[[2]])
sdm.names <- c("current", "future")
# Plot
sdm_leaflet(sdm = sdm.list,
            mode = "bin",
            pts = pa_data,
           layernames = sdm.names)

}
